requirements.txt
aiohttp
asgiref
async-timeout
attrs
beautifulsoup4
bs4
certifi
cffi
chardet
discord.py
Django
django-cors-headers
django-discover-runner
django-role-permissions
djangorestframework
idna
multidict
mutagen
psycopg2-binary
pycparser
PyNaCl
pytz
requests
six
soupsieve
sqlparse
urllib3
waitress
websockets
yarl
youtube-dl